Major rewrite in client-side javascript.
Have a gameObjectFactory() to create javascript objects.
Every javascript object which is visible is a GameObject and extends
from it.

Every object which implements GameObject should have a createContext()
method which returns a jQuery object for the just created container all
the stuff this object creates is put in. This container is then
available under this.$context. Also all members of the JSON-Data which
is retrieved from the server is automatically accessible as a member,
provided the superconstructor is called.